Baby are born with size relative to the parent determined by the baby genes. Bug fix: Physical properties of flask only take effect after first update. Other fine tuning.
Took 3 minutes

+++ Assets/Scripts/Environment/ChemicalBlob.cs
@@ -8,7 +8,7 @@ namespace Environment
     [RequireComponent(typeof(Rigidbody2D))]
     public class ChemicalBlob : PhysicalFlask
     {
-        public static readonly double MinBlobSize = .01f;
+        public static readonly float MinBlobSize = .01f;
 
         private void Start() => name = $"ChemicalBlob{GetInstanceID()}";
